requirements.txt
amqp==2.4.2
billiard==3.6.0.0
celery==4.3.0
Django==2.2
djangorestframework==3.9.2
kombu==4.5.0
pytz==2018.9
sqlparse==0.3.0
vine==1.3.0
